+++ protoc-gen-gomeet-service/main.go
@@ -50,6 +50,8 @@ func main() {
 		projectPkg      string
 		defaultPrefixes string
 		dbTypes         string
+		uiType          string
+		queueTypes      string
 		extraServeFlags string
 	)
 	useGogoImport := false
@@ -65,6 +67,10 @@ func main() {
 				subServices = strings.Split(parts[1], ",")
 			case "db_types":
 				dbTypes = parts[1]
+			case "ui_type":
+				uiType = parts[1]
+			case "queue_types":
+				queueTypes = parts[1]
 			case "default_prefixes":
 				defaultPrefixes = parts[1]
 			case "project_pkg":
@@ -96,6 +102,20 @@ func main() {
 		err := p.SetDbTypes(dbTypes)
 		if err != nil {
 			g.Error(err, "bad db_types parameter")
+		}
+	}
+
+	if uiType != "" {
+		err := p.SetUiType(uiType)
+		if err != nil {
+			g.Error(err, "bad ui_type parameter")
+		}
+	}
+
+	if queueTypes != "" {
+		err := p.SetQueueTypes(queueTypes)
+		if err != nil {
+			g.Error(err, "bad queue_types parameter")
 		}
 	}