+++ utils/project/project.go
@@ -22,16 +22,26 @@ import (
 	tmplHelpers "github.com/gomeet/gomeet/utils/project/templates/helpers"
 )
 
-const DEFAULT_PROTO_PKG_ALIAS = "pb"
+const (
+	DEFAULT_PROTO_PKG_ALIAS = "pb"
+	DEFAULT_ELM             = "elm-bulma"
+)
 
-var allowedDbTypes = []string{"mysql", "postgres", "postgis", "sqlite", "mssql"}
+var (
+	allowedDbTypes    = []string{"mysql", "postgres", "postgis", "sqlite", "mssql"}
+	allowedUiTypes    = []string{"none", "simple", "elm", "elm-bulma", "elm-milligram", "elm-minimal", "elm-minimal-http"} //TODO "react", "vuejs", ....
+	allowedQueueTypes = []string{"memory", "rabbitmq", "zeromq", "sqs"}
+)
 
 func GomeetDefaultPrefixes() string {
 	return helpers.GomeetDefaultPrefixes
 }
 
-func GomeetAllowedDbTypes() []string {
-	return allowedDbTypes
+func GomeetAllowedDbTypes() []string { return allowedDbTypes }
+func GomeetAllowedUiTypes() []string { return allowedUiTypes }
+
+func GomeetAllowedQueueTypes() []string {
+	return allowedQueueTypes
 }
 
 type serveFlag struct {
@@ -46,6 +56,8 @@ type Project struct {
 
 	SubServices          map[string]*Project
 	dbTypes              []string
+	uiType               string
+	queueTypes           []string
 	hasPostgis           bool
 	extraServeFlags      []*serveFlag
 	folder               *folder
@@ -67,7 +79,7 @@ func New(inputPath string) (*Project, error) {
 	if err != nil {
 		return nil, err
 	}
-	p := &Project{pkgNfo, nil, []string{}, false, nil, nil, nil, nil, nil, false, "0.0.1+dev"}
+	p := &Project{pkgNfo, nil, []string{}, "none", []string{}, false, nil, nil, nil, nil, nil, false, "0.0.1+dev"}
 	p.SetDefaultPrefixes("")
 	p.SetDefaultProtoPkgAlias("")
 	return p, nil
@@ -180,6 +192,51 @@ func (p *Project) SetDbTypes(s string) error {
 	return nil
 }
 
+func (p *Project) SetUiType(s string) error {
+	p.uiType = "none"
+	if s != "" {
+		uiType := strings.ToLower(strings.TrimSpace(s))
+		ok := false
+		for _, allowedUiType := range GomeetAllowedUiTypes() {
+			if uiType == allowedUiType {
+				ok = true
+				break
+			}
+		}
+		if !ok {
+			return fmt.Errorf("\"%s\" isn't allowed in ui_type - allowed ui_type : [%s]", uiType, strings.Join(GomeetAllowedUiTypes(), "|"))
+		}
+		if uiType == "elm" {
+			uiType = DEFAULT_ELM
+		}
+		p.uiType = uiType
+	}
+
+	return nil
+}
+
+func (p *Project) SetQueueTypes(s string) error {
+	if s != "" {
+		queueTypes := strings.Split(s, ",")
+		for _, queueType := range queueTypes {
+			queueType = strings.ToLower(strings.TrimSpace(queueType))
+			ok := false
+			for _, allowedQueueType := range GomeetAllowedQueueTypes() {
+				if queueType == allowedQueueType {
+					ok = true
+					break
+				}
+			}
+			if !ok {
+				return fmt.Errorf("%s isn't allowed queueType", queueType)
+			}
+			p.queueTypes = append(p.queueTypes, queueType)
+		}
+	}
+
+	return nil
+}
+
 func (p *Project) SetDefaultProtoPkgAlias(s string) error {
 	if s == "" {
 		s = DEFAULT_PROTO_PKG_ALIAS
@@ -197,6 +254,8 @@ func (p Project) GomeetGeneratorUrl() string                    { return "https:
 func (p Project) Version() string                               { return p.version }
 func (p Project) ProtoFiles() []*descriptor.FileDescriptorProto { return p.protoFiles }
 func (p Project) DbTypes() []string                             { return p.dbTypes }
+func (p Project) UiType() string                                { return p.uiType }
+func (p Project) QueueTypes() []string                          { return p.queueTypes }
 func (p Project) ExtraServeFlags() []*serveFlag                 { return p.extraServeFlags }
 
 func (p Project) GoCGOEnabled() int {
@@ -219,6 +278,33 @@ func (p Project) HasDb() bool {
 	}
 	for _, ss := range p.SubServices {
 		if ss.HasDb() {
+			return true
+		}
+	}
+
+	return false
+}
+
+func (p Project) HasUi() bool {
+	uiT := p.UiType()
+	return (uiT != "" && uiT != "none")
+}
+
+func (p Project) HasUiElm() bool {
+	uiT := p.UiType()
+	return strings.HasPrefix(uiT, "elm")
+}
+
+func (p Project) HasMemoryQueue() bool {
+	if len(p.QueueTypes()) > 0 {
+		for _, queueType := range p.QueueTypes() {
+			if queueType == "memory" {
+				return true
+			}
+		}
+	}
+	for _, ss := range p.SubServices {
+		if ss.HasMemoryQueue() {
 			return true
 		}
 	}
@@ -318,6 +404,33 @@ func (p *Project) SubServicesMonolithHelp() string {
 				)
 			}
 		}
+		if len(ss.QueueTypes()) > 0 {
+			for _, queueType := range ss.QueueTypes() {
+				switch queueType {
+				case "memory":
+					ssFlags = append(
+						ssFlags,
+						fmt.Sprintf(
+							"\"svc-%s-%s-queue-worker-count\"",
+							ss.ShortName(),
+							strings.ToLower(queueType),
+						),
+					)
+					ssFlags = append(
+						ssFlags,
+						fmt.Sprintf(
+							"\"svc-%s-%s-queue-max-size\"",
+							ss.ShortName(),
+							strings.ToLower(queueType),
+						),
+					)
+				case "rabbitmq", "zeromq", "sqs":
+				// rabbitmq, zeromq, sqs support is not yet implemented
+				default:
+					// FIXME what to do?
+				}
+			}
+		}
 		if len(ss.ExtraServeFlags()) > 0 {
 			for _, ssf := range ss.ExtraServeFlags() {
 				ssFlags = append(
@@ -335,7 +448,7 @@ func (p *Project) SubServicesMonolithHelp() string {
 			ssString,
 			strings.Join(ssFlags, ", "),
 		)
-		if len(ss.DbTypes()) > 0 || len(ss.ExtraServeFlags()) > 0 {
+		if len(ss.DbTypes()) > 0 || len(ss.QueueTypes()) > 0 || len(ss.ExtraServeFlags()) > 0 {
 			ssString = fmt.Sprintf(
 				"%s flags are used to launch \"svc-%s\" server in the main process",
 				ssString,
@@ -373,6 +486,13 @@ func (p *Project) SubServicesDef() string {
 				"%s[db_types@%s]",
 				ssString,
 				strings.Join(ss.DbTypes(), "|"),
+			)
+		}
+		if len(ss.QueueTypes()) > 0 {
+			ssString = fmt.Sprintf(
+				"%s[queue_types@%s]",
+				ssString,
+				strings.Join(ss.QueueTypes(), "|"),
 			)
 		}
 		if len(ss.SubServices) > 0 {
@@ -428,6 +548,7 @@ func (p *Project) SetSubServices(subServices []string) error {
 			var (
 				ssVersion     string
 				ssDbTypes     []string
+				ssQueueTypes  []string
 				ssFlags       []string
 				ssSubServices []string
 			)
@@ -435,6 +556,8 @@ func (p *Project) SetSubServices(subServices []string) error {
 				switch {
 				case strings.HasPrefix(ssFlag, "db_types@"):
 					ssDbTypes = strings.Split(strings.TrimPrefix(ssFlag, "db_types@"), "|")
+				case strings.HasPrefix(ssFlag, "queue_types@"):
+					ssQueueTypes = strings.Split(strings.TrimPrefix(ssFlag, "queue_types@"), "|")
 				case strings.HasPrefix(ssFlag, "version@"):
 					ssVersion = strings.TrimPrefix(ssFlag, "version@")
 				case strings.HasPrefix(ssFlag, "sub_services@"):
@@ -452,6 +575,9 @@ func (p *Project) SetSubServices(subServices []string) error {
 			}
 			if len(ssDbTypes) > 0 {
 				subSvc.SetDbTypes(strings.Join(ssDbTypes, ","))
+			}
+			if len(ssQueueTypes) > 0 {
+				subSvc.SetQueueTypes(strings.Join(ssQueueTypes, ","))
 			}
 			if len(ssSubServices) > 0 {
 				neededSubServices = append(neededSubServices, ssSubServices...)
@@ -535,6 +661,14 @@ func (p *Project) setProjectCreationTree(keepFile, keepProtoModel bool) (err err
 		f.delete("third_party/github.com/gogo")
 	}
 
+	f.delete("ui")
+	if p.HasUi() {
+		_, err := f.addTree("ui", fmt.Sprintf("project-creation/ui/%s", p.UiType()), nil, keepFile)
+		if err != nil {
+			return err
+		}
+	}
+
 	p.folder = f
 
 	return nil
@@ -576,6 +710,14 @@ func (p Project) AfterProjectCreationCmd() (r []string) {
 	r = append(r, "make tools-sync proto dep")
 	r = append(r, "git add .")
 	r = append(r, "git commit -m 'Added tools and dependencies'")
+	switch {
+	case p.HasUi() && p.HasUiElm():
+		r = append(r, "make ui-setup")
+		r = append(r, "make ui")
+		r = append(r, "git add .")
+		r = append(r, "git commit -m 'Added ui'")
+	}
+
 	return r
 }
 
@@ -714,6 +856,7 @@ func (p *Project) GenFromProto(req *plugin.CodeGeneratorRequest) error {
 	svc.addFile("service_test.go", "protoc-gen/service/service_test.go.tmpl", nil, false)
 	svc.addFile("init_subservice_clients.go", "protoc-gen/service/init_subservice_clients.go.tmpl", nil, false)
 	svc.addFile("init_databases.go", "protoc-gen/service/init_databases.go.tmpl", nil, false)
+	svc.addFile("init_queues.go", "protoc-gen/service/init_queues.go.tmpl", nil, false)
 	f.addTree("infra", "protoc-gen/infra", nil, false)
 	f.addTree("hack", "protoc-gen/hack", nil, false)
 	protoPkg, err := p.GoProtoPkgAlias()